+++ docs/examples/daterange/mobile.tsx
@@ -10,14 +10,14 @@ export default function Example() {
     <DeviceTypeProvider deviceType="mobile">
       <Flex alignItems="center" height="100%" justifyContent="center" width="100%">
         <DateRange
-          endDateValue={endDate}
+          dateValue={{ startDate, endDate }}
           onCancel={() => {}}
-          onEndDateChange={({ value }) => setEndDate(value)}
-          onEndDateError={() => {}}
-          onStartDateChange={({ value }) => setStartDate(value)}
-          onStartDateError={() => {}}
+          onDateChange={(newStartDate, newEndDate) => {
+            setStartDate(newStartDate.value);
+            setEndDate(newEndDate.value);
+          }}
+          onDateError={{ startDate: () => {}, endDate: () => {} }}
           onSubmit={() => {}}
-          startDateValue={startDate}
         />
       </Flex>
     </DeviceTypeProvider>